#46: Extracted putty related menu from MainForm.
