+++ Source/Terminals/Forms/MainForm.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Drawing;
-using System.IO;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using System.Windows.Forms;
@@ -130,8 +129,6 @@ namespace Terminals
                 this.Font = SystemFonts.IconTitleFont;
 
                 InitializeComponent(); // main designer procedure
-
-                InitializeToolsMenu();
 
                 this.formSettings = new FormSettings(this);
                 this.tabsFilter = new TabControlFilter(this.tcTerminals);
@@ -1400,48 +1397,5 @@ namespace Terminals
         {
             this.tabControlRemover.Disconnect();
         }
-
-        #region Putty to move
-
-        private void InitializeToolsMenu()
-        {
-            ToolStripMenuItem openSshAgentToolStripMenuItem = new ToolStripMenuItem();
-            openSshAgentToolStripMenuItem.Name = "openSshAgentToolStripMenuItem";
-            openSshAgentToolStripMenuItem.Size = new Size(276, 22);
-            openSshAgentToolStripMenuItem.Text = "Open SSH Agent";
-            openSshAgentToolStripMenuItem.ToolTipText = "Open SSH Agent";
-            openSshAgentToolStripMenuItem.Click += new EventHandler(this.OpenSshAgentToolStripMenuItem_Click);
-            int separatorIndex = this.toolsToolStripMenuItem.DropDownItems.IndexOf(this.toolStripSeparator9);
-            this.toolsToolStripMenuItem.DropDownItems.Insert(separatorIndex, openSshAgentToolStripMenuItem);
-
-            ToolStripMenuItem openSshKeygenToolStripMenuItem = new ToolStripMenuItem();
-            openSshKeygenToolStripMenuItem.Name = "openSshKeygenToolStripMenuItem";
-            openSshKeygenToolStripMenuItem.Size = new Size(276, 22);
-            openSshKeygenToolStripMenuItem.Text = "Open SSH Keygen";
-            openSshKeygenToolStripMenuItem.ToolTipText = "Open SSH Keygen";
-            openSshKeygenToolStripMenuItem.Click += new EventHandler(this.OpenSshKeygenToolStripMenuItem_Click);
-            separatorIndex = this.toolsToolStripMenuItem.DropDownItems.IndexOf(this.toolStripSeparator9);
-            this.toolsToolStripMenuItem.DropDownItems.Insert(separatorIndex, openSshKeygenToolStripMenuItem);
-
-        }
-
-        // todo Move openig putty tools to putty plugin as menu extender
-        private void OpenSshAgentToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            this.OpenPuttyTool("pageant.exe");
-        }
-
-        private void OpenSshKeygenToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            this.OpenPuttyTool("puttygen.exe");
-        }
-
-        private void OpenPuttyTool(string name)
-        {
-            string path = Path.Combine(PluginsLoader.FindBasePluginDirectory(), "Putty", "Resources", name);
-            ExternalLinks.OpenPath(path);
-        }
-
-        #endregion
     }
 }
