+++ Source/Terminals/Terminals.csproj
@@ -189,6 +189,7 @@
     <Compile Include="Connections\DesktopShares.cs" />
     <Compile Include="Connections\PluginDefinition.cs" />
     <Compile Include="Forms\IConnectionCommands.cs" />
+    <Compile Include="Forms\PuttyMenuVisitor.cs" />
     <Compile Include="IStartupUi.cs" />
     <Compile Include="Network\PortScanner\NetworkAdapters.cs" />
     <Compile Include="Services\ExternalLinks.cs" />
